requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.78
botocore==1.35.78
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
jmespath==1.0.1
kombu==5.4.2
moviepy==2.1.1
numpy==2.2.0
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
